wip(#1102): extract emit/locals.rs
